+++ src/locale.js
@@ -187,7 +187,7 @@ SnapTranslator.dict.de = {
     'translator_e-mail':
         'jens@moenig.org, jadga.huegle@sap.com',
     'last_changed':
-        '2022-06-22'
+        '2022-06-23'
 };
 
 SnapTranslator.dict.it = {