+++ HISTORY.md
@@ -54,6 +54,7 @@
 
 ### 2022-06-23
 * objects, threads: turned DEFINE into a command block with an upvar, experimental
+* German translation update for the new format of the DEFINE block
 
 ### 2022-06-22
 * objects, blocks, threads: new "this script" primitive in sensing, experimental